crawler_projects/requirements.txt
scrapy
redis
httpx
BeautifulSoup4
scrapy_redis
scrapyd
Pillow
python-dotenv
